+++ Sources/Client/Client/Client+Setup.swift
@@ -48,7 +48,7 @@ extension Client {
         }
         
         userAtomic.set(user)
-        urlSession = setupURLSession()
+        urlSession = makeURLSession()
         logger?.log("Sending a request for a Guest Token...")
         
         request(endpoint: .guestToken(user)) { [unowned self] (result: Result<TokenResponse, ClientError>) in
@@ -179,10 +179,8 @@ extension Client {
         let completion = completion == nil ? nil : { result in DispatchQueue.main.async { completion?(result) } }
         
         do {
-            let webSocket = try setupWebSocket(user: user, token: token)
-            database?.user = user
-            self.webSocket = webSocket
-            urlSession = setupURLSession(token: token)
+            webSocket.request = try makeWebSocketRequest(user: user, token: token)
+            urlSession = makeURLSession(token: token)
             self.token = token
             
             if let completion = completion {