+++ Tests/Client/Unit Tests/WebSocket/WebSocketProviderMock.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import StreamChatClient
 
-class WebSocketProviderMock: WebSocketProvider {
+final class WebSocketProviderMock: WebSocketProvider {
     var request: URLRequest
     
     var isConnected: Bool = false
@@ -24,7 +24,7 @@ class WebSocketProviderMock: WebSocketProvider {
     var sendPingCalledCounter = 0
 
     convenience init() {
-        self.init(request: .init(url: URL(string: "test_url")!), callbackQueue: .main)
+        self.init(request: .init(url: .placeholder), callbackQueue: .main)
     }
     
     required init(request: URLRequest, callbackQueue: DispatchQueue) {
@@ -37,6 +37,8 @@ class WebSocketProviderMock: WebSocketProvider {
     }
     
     func disconnect() {
+        isConnected = false
+        delegate?.websocketDidDisconnect(error: nil)
     }
     
     func sendPing() {